--- __init__.py
@@ -1,2 +1,0 @@
-from .python.app.console import ShotgunPythonConsoleWidget, PythonConsoleWidget
-
